backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
opencv-python==4.11.0.86
numpy==2.2.5
Pillow==11.2.1
